Customizer theme switcher: Fix some esoteric breakage in iOS Safari.
This manifested in only being able to switch your theme once in the customizer before the "Change" button would appear to stop responding, though a switch to "Preview" and back to "Customize" would reveal the theme switcher after all. Animated positioning appears to be the problem here. The animations are not critical to the experience, particularly on touch+smaller screen devices.

props iseulde.
see #31794.